refactor: remove settings menu
