+++ src/components/messenger/user-profile/overview-panel/rewards-item/container.test.tsx
@@ -1,4 +1,4 @@
-import { StoreBuilder } from '../../../store/test/store';
+import { StoreBuilder } from '../../../../../store/test/store';
 import { Container } from './container';
 
 describe(Container, () => {
