+++ src/components/messenger/user-profile/overview-panel/rewards-item/index.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-import { bemClassName } from '../../../lib/bem';
+import { bemClassName } from '../../../../../lib/bem';
 import './styles.scss';
 
 const cn = bemClassName('rewards-item');
