+++ src/components/messenger/user-profile/overview-panel/rewards-item/styles.scss
@@ -1,5 +1,5 @@
 @use '~@zero-tech/zui/styles/theme' as theme;
-@import '../../../glass';
+@import '../../../../../glass';
 
 .rewards-item {
   display: flex;
