--- src/components/settings-menu/styles.scss
@@ -1,68 +1,0 @@
-@use '~@zero-tech/zui/styles/theme' as theme;
-
-.settings-menu {
-  max-width: 195px;
-
-  > *:first-child {
-    &:hover {
-      background: none;
-      cursor: default;
-    }
-  }
-
-  .header {
-    display: flex;
-    align-items: center;
-    gap: 16px;
-    pointer-events: none;
-    overflow: hidden;
-
-    &__user-details {
-      display: flex;
-      flex-direction: column;
-      gap: 2px;
-      overflow: hidden;
-    }
-
-    &__name {
-      text-overflow: ellipsis;
-      overflow: hidden;
-      white-space: nowrap;
-
-      font-size: 14px;
-      font-weight: 700;
-      line-height: 17px;
-      color: theme.$color-greyscale-12;
-    }
-
-    &__handle {
-      font-size: 10px;
-      line-height: 12px;
-      color: theme.$color-greyscale-11;
-
-      span {
-        color: theme.$color-greyscale-11;
-      }
-    }
-  }
-
-  .option {
-    gap: 10px;
-  }
-
-  .zui-dropdown-item.divider {
-    border-bottom: 1px solid rgba(163, 162, 163, 0.1);
-    margin: 0;
-    padding: 0;
-  }
-
-  .zui-dropdown-item.logout {
-    color: theme.$color-failure-11;
-  }
-
-  &__secure-backup-modal {
-    &:focus-visible {
-      outline: none;
-    }
-  }
-}
